requirements.txt
flask
pandas
numpy
scikit-learn
joblib
python-dotenv
requests
matplotlib
seaborn
google-generativeai